fix(recordings): keep event history within size gate
