+++ pkg/services/recordings/internal/events/event_history.go
@@ -990,13 +990,3 @@ func uniqueStrings(input []string) []string {
 	}
 	return out
 }
-
-func cloneExpectedArtifactTemplateContext(
-	context *work.ExpectedArtifactTemplateContext,
-) *work.ExpectedArtifactTemplateContext {
-	if context == nil {
-		return nil
-	}
-	clone := *context
-	return &clone
-}
